Refactor assistant context into notebook contribution
Extract the assistant context gathering logic from PositronNotebookInstance
into a dedicated PositronNotebookAssistantController contribution. This follows
the existing notebook contribution pattern and keeps the instance class cleaner.
